[MCJ-1489] refactor: 마이페이지 참여 응답 매핑 로직을 DTO from으로 일원화
