+++ src/main/kotlin/com/moongchijang/domain/participation/application/dto/InProgressParticipationItemResponse.kt
@@ -1,7 +1,10 @@
 package com.moongchijang.domain.participation.application.dto
 
+import com.moongchijang.domain.groupbuy.application.dto.GroupBuyProgressCalculator
+import com.moongchijang.domain.participation.domain.entity.Participation
 import io.swagger.v3.oas.annotations.media.Schema
 import java.time.LocalDateTime
+import java.time.temporal.ChronoUnit
 
 @Schema(description = "마이페이지 진행 중 탭 참여 카드 아이템")
 data class InProgressParticipationItemResponse(
@@ -35,4 +38,32 @@ data class InProgressParticipationItemResponse(
 
     @field:Schema(description = "참여 일시", example = "2026-04-12T10:30:00")
     val participatedAt: LocalDateTime
-)
+) {
+    companion object {
+        fun from(participation: Participation, now: LocalDateTime = LocalDateTime.now()): InProgressParticipationItemResponse {
+            val groupBuy = participation.groupBuy
+            val dDay = ChronoUnit.DAYS.between(
+                now.toLocalDate(),
+                groupBuy.deadline.toLocalDate()
+            ).toInt()
+
+            return InProgressParticipationItemResponse(
+                participationId = participation.id,
+                groupBuyId = groupBuy.id,
+                productName = groupBuy.productName,
+                storeName = groupBuy.store.name,
+                pickupAt = LocalDateTime.of(groupBuy.pickupDate, groupBuy.pickupTimeStart),
+                paidAmount = participation.totalAmount,
+                quantity = participation.quantity,
+                achievementRate = GroupBuyProgressCalculator.achievementRate(
+                    currentQuantity = groupBuy.currentQuantity,
+                    targetQuantity = groupBuy.targetQuantity
+                ),
+                dDay = dDay,
+                participatedAt = requireNotNull(participation.createdAt) {
+                    "[InProgressParticipationItemResponse] 참여 생성일시 누락: participationId=${participation.id}"
+                }
+            )
+        }
+    }
+}
